feat: silent give command
