+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/commands/subcommands/AbstractGiveCommand.java
@@ -1,0 +1,109 @@
+package io.github.thebusybiscuit.slimefun4.core.commands.subcommands;
+
+import java.util.Locale;
+import java.util.Map;
+import java.util.Optional;
+
+import javax.annotation.ParametersAreNonnullByDefault;
+
+import org.bukkit.command.CommandSender;
+import org.bukkit.entity.Player;
+import org.bukkit.inventory.ItemStack;
+
+import io.github.bakedlibs.dough.common.CommonPatterns;
+import io.github.bakedlibs.dough.common.PlayerList;
+import io.github.bakedlibs.dough.items.CustomItemStack;
+import io.github.thebusybiscuit.slimefun4.api.items.SlimefunItem;
+import io.github.thebusybiscuit.slimefun4.core.commands.SlimefunCommand;
+import io.github.thebusybiscuit.slimefun4.core.commands.SubCommand;
+import io.github.thebusybiscuit.slimefun4.core.multiblocks.MultiBlockMachine;
+import io.github.thebusybiscuit.slimefun4.implementation.Slimefun;
+
+/**
+ * This is an abstract command that is used to cheat items to players.
+ * Extending class decides whether the receiver is notified or not.
+ *
+ * @author TheBusyBiscuit
+ * @author ybw0014
+ */
+abstract class AbstractGiveCommand extends SubCommand {
+
+    private static final String PLACEHOLDER_PLAYER = "%player%";
+    private static final String PLACEHOLDER_ITEM = "%item%";
+    private static final String PLACEHOLDER_AMOUNT = "%amount%";
+
+    private final boolean silent;
+
+    @ParametersAreNonnullByDefault
+    AbstractGiveCommand(Slimefun plugin, SlimefunCommand cmd, String name, boolean silent) {
+        super(plugin, cmd, name, false);
+
+        this.silent = silent;
+    }
+
+    @Override
+    public void onExecute(CommandSender sender, String[] args) {
+        if (sender.hasPermission("slimefun.cheat.items") || !(sender instanceof Player)) {
+            if (args.length > 2) {
+                Optional<Player> player = PlayerList.findByName(args[1]);
+
+                if (player.isPresent()) {
+                    Player p = player.get();
+
+                    SlimefunItem sfItem = SlimefunItem.getById(args[2].toUpperCase(Locale.ROOT));
+
+                    if (sfItem != null) {
+                        giveItem(sender, p, sfItem, args);
+                    } else {
+                        Slimefun.getLocalization().sendMessage(sender, "messages.invalid-item", true, msg -> msg.replace(PLACEHOLDER_ITEM, args[2]));
+                    }
+                } else {
+                    Slimefun.getLocalization().sendMessage(sender, "messages.not-online", true, msg -> msg.replace(PLACEHOLDER_PLAYER, args[1]));
+                }
+            } else {
+                Slimefun.getLocalization().sendMessage(sender, "messages.usage", true, msg -> msg.replace("%usage%", "/sf " + getName() + " <Player> <Slimefun Item> [Amount]"));
+            }
+        } else {
+            Slimefun.getLocalization().sendMessage(sender, "messages.no-permission", true);
+        }
+    }
+
+    private void giveItem(CommandSender sender, Player p, SlimefunItem sfItem, String[] args) {
+        if (sfItem instanceof MultiBlockMachine) {
+            Slimefun.getLocalization().sendMessage(sender, "guide.cheat.no-multiblocks");
+        } else {
+            int amount = parseAmount(args);
+
+            if (amount > 0) {
+                if (!silent) {
+                    Slimefun.getLocalization().sendMessage(p, "messages.given-item", true, msg -> msg.replace(PLACEHOLDER_ITEM, sfItem.getItemName()).replace(PLACEHOLDER_AMOUNT, String.valueOf(amount)));
+                }
+                Map<Integer, ItemStack> excess = p.getInventory().addItem(new CustomItemStack(sfItem.getItem(), amount));
+                if (Slimefun.getCfg().getBoolean("options.drop-excess-sf-give-items") && !excess.isEmpty()) {
+                    for (ItemStack is : excess.values()) {
+                        p.getWorld().dropItem(p.getLocation(), is);
+                    }
+                }
+
+                Slimefun.getLocalization().sendMessage(sender, "messages.give-item", true, msg -> msg.replace(PLACEHOLDER_PLAYER, args[1]).replace(PLACEHOLDER_ITEM, sfItem.getItemName()).replace(PLACEHOLDER_AMOUNT, String.valueOf(amount)));
+            } else {
+                Slimefun.getLocalization().sendMessage(sender, "messages.invalid-amount", true, msg -> msg.replace(PLACEHOLDER_AMOUNT, args[3]));
+            }
+        }
+    }
+
+    private int parseAmount(String[] args) {
+        int amount = 1;
+
+        if (args.length == 4) {
+            if (CommonPatterns.NUMERIC.matcher(args[3]).matches()) {
+                amount = Integer.parseInt(args[3]);
+            } else {
+                return 0;
+            }
+        }
+
+        return amount;
+    }
+
+}
